moved Ui.java to ui package
