+++ src/main/java/dolla/command/AddCommand.java
@@ -1,7 +1,7 @@
 package dolla.command;
 
 import dolla.Time;
-import dolla.Ui;
+import dolla.ui.Ui;
 import dolla.task.Task;
 import dolla.task.TaskList;
