+++ src/main/java/dolla/command/AddDebtsCommand.java
@@ -1,7 +1,7 @@
 package dolla.command;
 
 import dolla.DollaData;
-import dolla.Ui;
+import dolla.ui.Ui;
 import dolla.action.Redo;
 import dolla.action.Undo;
 import dolla.task.Debt;
