+++ src/main/java/dolla/command/RemoveCommand.java
@@ -1,7 +1,7 @@
 package dolla.command;
 
 import dolla.DollaData;
-import dolla.Ui;
+import dolla.ui.Ui;
 import dolla.action.Redo;
 import dolla.action.Undo;
 import dolla.task.RecordList;
@@ -57,7 +57,6 @@ public class RemoveCommand extends Command {
             Undo.addCommand(mode, recordList.get().get(logNumInt).getUserInput(), logNumInt);
             Redo.clearRedo(mode);
         }
-
         Ui.echoRemove(recordList.get().get(logNumInt).getRecordDetail());
         dollaData.removeFromRecordList(mode,logNumInt);
     }
