+++ src/main/java/dolla/parser/Parser.java
@@ -1,14 +1,11 @@
 package dolla.parser;
 
 import dolla.Time;
-import dolla.Ui;
-import dolla.command.AddDeadlineCommand;
-import dolla.command.AddEventCommand;
+import dolla.ui.Ui;
 import dolla.command.Command;
 import dolla.command.ErrorCommand;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 
 /**
