+++ src/main/java/dolla/sort/SortDescription.java
@@ -1,6 +1,6 @@
 package dolla.sort;
 
-import dolla.Ui;
+import dolla.ui.Ui;
 import dolla.task.Record;
 
 import java.util.ArrayList;
