+++ src/main/java/dolla/sort/SortName.java
@@ -1,6 +1,6 @@
 package dolla.sort;
 
-import dolla.Ui;
+import dolla.ui.Ui;
 import dolla.task.Record;
 
 import java.util.ArrayList;
